auth cleaned up
