+++ src/server/auth/withSession.ts
@@ -2,6 +2,7 @@ import type { Session, User } from "next-auth";
 import type { RequestEventLoader } from "~/server/types";
 import { paths } from "~/utils/paths";
 import { prisma } from "../db/client";
+import { getServerSession } from "./nextAuth";
 import { authOptions } from "./options";
 
 export type ProtectedRequestContext = {
@@ -20,7 +21,7 @@ export const withProtectedSession = <
   options: WithProtectedSessionOptions = {}
 ) => {
   return async (event: R) => {
-    const session = await getSession(event, authOptions);
+    const session = await getServerSession(event, authOptions);
 
     if (!session || !session.user) {
       throw event.redirect(302, options.redirectTo || paths.signIn);
